requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.2
djangorestframework==3.15.2
filelock==3.16.1
platformdirs==4.3.6
psycopg==3.2.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.27.0
whitenoise==6.7.0
gunicorn
